requirements.txt
flask==3.0.0
flask-cors==4.0.0
pandas==2.1.4
openpyxl==3.1.2
openai==0.28.0
python-dotenv==1.0.0
werkzeug==3.0.1
azure-devops==7.1.0b4